streamlitApp/requirements.txt
openai==0.27.4
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.1
streamlit==1.21.0
